refactor: rename nameConversions.js -> stringConversions.js
